+++ yecs/move.go
@@ -3,8 +3,8 @@ package yecs
 import "yam/y3d"
 
 type Move struct {
-	AnglularSpeed float64
-	ForwardSpeed  float64
+	AnglularSpeed float32
+	ForwardSpeed  float32
 }
 
 type MoveSystem2D struct{}
@@ -18,14 +18,14 @@ func (ms MoveSystem2D) Update(w *World, dt float64, entites []EntityId) {
 		move := w.GetComponent(e, MoveComponent).(Move)
 		transform := w.GetComponent(e, TransformComponent).(Transform)
 		if move.AnglularSpeed > y3d.NearZero {
-			angle := move.AnglularSpeed * dt
-			inc := y3d.FromAngleAxis(y3d.UNIT_Z, angle)
+			angle := move.AnglularSpeed * float32(dt)
+			inc := y3d.FromAngleAxis(y3d.UNIT_Z, float64(angle))
 			transform.Orientation = y3d.ProdQuaternion(inc, transform.Orientation)
 			transform.NeedCalculation = true
 		}
 		if move.ForwardSpeed > y3d.NearZero {
 			velocity := y3d.Smul(transform.GetForward(), move.ForwardSpeed)
-			transform.Position = y3d.Add(transform.Position, y3d.Smul(velocity, dt))
+			transform.Position = y3d.Add(transform.Position, y3d.Smul(velocity, float32(dt)))
 			transform.NeedCalculation = true
 		}
 		if transform.NeedCalculation {
